backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.20
pyjwt==2.8.0
bcrypt==4.0.1
httpx==0.24.1
python-dotenv==1.0.0
pytest==7.4.0
langchain==0.0.300